src/WebApp/pyproject.toml
[tool.poetry]
name = "webapp"
version = "0.1.0"
description = ""
authors = ["lxstvayne <ssavi2002@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.5"
django = "^4.1.7"
django-phone-field = "^1.8.1"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.25.1"
django-cors-headers = "^3.13.0"
ipython = "^8.10.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
